split artifacts from lens
